Adding text labels to the examples for sanity
Adding text labels to the examples because a bunch of images of cubes
and spheres are really hard to understand, especially when you want to
quickly come to the homepage and find an example, but it's all images,
and even the example pages themselves don't have labels :( Also updates
CSS for minor styling alignment.

However, it looks like something bad is going on here? The examples on
master don't line up with the examples on the homepage. Why the
discrepancy? It looks like the branch `gh-pages` has the up to date
examples...(?) I can make the PR to that branch as well